Rename mouse module to pointer
